feat: 7주차 과제1번

+++ likelion_11th/main/views.py
@@ -1,14 +1,29 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import Post
+from .models import Post, Comment, Tag
 from django.utils import timezone
-from django.contrib.auth.decorators import login_required
+import re
+# from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
 def detail(request, id):
     post = get_object_or_404(Post, pk=id)
-    is_author = request.user.is_authenticated and post.writer == request.user.profile.nickname
-    return render(request, "main/detail.html", {"post": post, 'is_author':is_author})
+    if request.method == 'GET':
+        comments = Comment.objects.filter(post=post)
+        return render(request, "main/detail.html",{
+            'post':post,
+            'comments':comments,
+        })
+    elif request.method == 'POST':
+        new_comment = Comment()
+        # 댓글에 post 객체 넣기
+        new_comment.post = post
+        # 댓글 작성자 객체 넣기
+        new_comment.writer = request.user
+        new_comment.content = request.POST['content']
+        new_comment.pub_date = timezone.now()
+        new_comment.save()
+        return redirect("main/detail.html", id)
 
 
 
@@ -26,38 +41,67 @@ def new(request):
 
 
 def create(request):
-    new_post = Post()
-    new_post.title = request.POST["title"]
-    new_post.writer = request.POST["writer"]
-    new_post.pub_date = timezone.now()
-    new_post.image = request.FILES.get("image")
-    new_post.body = request.POST["body"]
-    new_post.mood = request.POST["mood"]
-    new_post.tmi = request.POST["tmi"]
-    new_post.save()
+    if request.user.is_authenticated:
+        new_post = Post()
+        new_post.title = request.POST["title"]
+        new_post.writer = request.user
+        new_post.pub_date = timezone.now()
+        new_post.image = request.FILES.get("image")
+        new_post.body = request.POST["body"]
+        new_post.mood = request.POST["mood"]
+        new_post.tmi = request.POST["tmi"]
+        new_post.save()
+        # 정규식 사용 #가 붙은 단어그룹 찾아서 리스트로 반환한 것을 tag_list에 넣음.
+        regex = re.compile('#([^\s`~!@#$%^&*()-+=\\\[\]\{\},.<>?]+)')
+        tag_list = re.findall(regex, new_post.body)
+        for tags in tag_list:
+            # 기존 get, 없으면 create
+            tag, boolean = Tag.objects.get_or_create(name=tags)
+            # 이후 필드에 추가
+            new_post.tags.add(tag.id)
+        return redirect("main:detail", new_post.id)
+    else:
+        return redirect('accounts:login')
 
-    return redirect("main:detail", new_post.id)
 
-@login_required
 def edit(request, id):
     edit_post = Post.objects.get(id=id)
     return render(request, "main/edit.html", {"post": edit_post})
 
-@login_required
-def update(request, id):
-    update_post = Post.objects.get(id=id)
-    update_post.title = request.POST["title"]
-    update_post.writer = request.POST["writer"]
-    update_post.pub_date = timezone.now()
-    update_post.image = request.FILES.get("image",update_post.image)
-    update_post.body = request.POST["body"]
-    update_post.mood = request.POST["mood"]
-    update_post.tmi = request.POST["tmi"]
-    update_post.save()
-    return redirect("main:detail", update_post.id)
 
-@login_required
+def update(request, id):
+    if request.user.is_authenticated:
+        update_post = Post.objects.get(id=id)
+        if request.user == update_post.writer:
+            update_post.title = request.POST["title"]
+            update_post.pub_date = timezone.now()
+            update_post.image = request.FILES.get("image",update_post.image)
+            update_post.body = request.POST["body"]
+            update_post.mood = request.POST["mood"]
+            update_post.tmi = request.POST["tmi"]
+            update_post.save()
+            return redirect("main:detail", update_post.id)
+    return redirect("accounts:login")
+
+
 def delete(request, id):
-    delete_post = Post.objects.get(id=id)
-    delete_post.delete()
-    return redirect("main:mainpage")
+    if request.user.is_authenticated:
+        delete_post = Post.objects.get(id=id)
+        delete_post.delete()
+        return redirect("main:mainpage")
+    
+    return redirect("accounts:login")
+
+def tag_list(request):
+    tags = Tag.objects.all()
+    return render(request,'main/tag_list.html',{
+        'tags' : tags,
+    })
+
+def tag_posts(request, tag_id):
+    tag = get_object_or_404(Tag, id = tag_id)
+    posts = Post.objects.all()
+    return render(request, 'main/tag_posts.html',{
+        'tag':tag,
+        'posts':posts,
+    })
